Rename gulrak filesystem.hh to .hpp to mirror upstream
