+++ src/phase1.hpp
@@ -38,7 +38,7 @@
 #define _HAS_STD_BYTE 0
 #define NOMINMAX
 
-#include "../lib/include/filesystem.hh"
+#include "../lib/include/filesystem.hpp"
 namespace fs = ghc::filesystem;
 #include "calculate_bucket.hpp"
 #include "entry_sizes.hpp"
