+++ src/plotter_disk.hpp
@@ -35,7 +35,7 @@
 #define _HAS_STD_BYTE 0
 #define NOMINMAX
 
-#include "../lib/include/filesystem.hh"
+#include "../lib/include/filesystem.hpp"
 namespace fs = ghc::filesystem;
 #include "calculate_bucket.hpp"
 #include "encoding.hpp"
